# backend/src/shared/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Controlador Financeiro — API de Gestão de CT
  description: API para gestão de Centros de Treinamento de Artes Marciais. Multi-tenant via account_id extraído do JWT.
  version: 1.0.0
  contact:
    name: Suporte
servers:
  - url: http://localhost:3000
    description: Desenvolvimento local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Erro:
      type: object
      properties:
        erro:
          type: object
          properties:
            mensagem:
              type: string
              example: Recurso não encontrado
            codigo:
              type: string
              example: NOT_FOUND
            status:
              type: integer
              example: 404
            requestId:
              type: string
              format: uuid
              example: 550e8400-e29b-41d4-a716-446655440000
    Meta:
      type: object
      properties:
        pagina:
          type: integer
          example: 1
        limite:
          type: integer
          example: 20
        total:
          type: integer
          example: 42
        totalPaginas:
          type: integer
          example: 3

    # ---------- AUTH ----------
...
    CtRequest:
      type: object
      required: [nome]
      properties:
        nome:
          type: string
          example: CT Fight Academy
        ativo:
          type: boolean
    Ct:
      type: object
      properties:
        id:
          type: integer
        account_id:
          type: integer
        nome:
          type: string
        ativo:
          type: integer
        criado_em:
          type: string
          format: date-time

    # ---------- ALUNO ----------
...
    ProfissionalRequest:
      type: object
      required: [nome]
      properties:
        nome:
          type: string
        email:
          type: string
          format: email
          nullable: true
        telefone:
          type: string
          nullable: true
        especialidade:
          type: string
          nullable: true
    Profissional:
      type: object
      properties:
        id:
          type: integer
        account_id:
          type: integer
        nome:
          type: string
        email:
          type: string
          nullable: true
        telefone:
          type: string
          nullable: true
        especialidade:
          type: string
          nullable: true
        ativo:
          type: integer

    # ---------- MODALIDADE ----------
    ModalidadeRequest:
      type: object
      required: [nome]
      properties:
        nome:
          type: string
        descricao:
          type: string
          nullable: true
    Modalidade:
      type: object
      properties:
        id:
          type: integer
        account_id:
          type: integer
        nome:
          type: string
        descricao:
          type: string
          nullable: true
        ativo:
          type: integer

    # ---------- HORÁRIO DE AULA ----------
    HorarioAulaRequest:
      type: object
      required: [ct_id, profissional_id, modalidade_id, dia_semana, hora_inicio, hora_fim]
      properties:
        ct_id:
          type: integer
        profissional_id:
          type: integer
        modalidade_id:
          type: integer
        dia_semana:
          type: integer
          minimum: 0
          maximum: 6
        hora_inicio:
          type: string
          example: '08:00'
        hora_fim:
          type: string
          example: '09:00'
        limite_vagas:
          type: integer
          nullable: true
    HorarioAula:
      type: object
      properties:
        id:
          type: integer
        account_id:
          type: integer
        ct_id:
          type: integer
        profissional_id:
          type: integer
        modalidade_id:
          type: integer
        dia_semana:
          type: integer
        hora_inicio:
          type: string
        hora_fim:
          type: string
        limite_vagas:
          type: integer
          nullable: true
        ativo:
          type: integer

    # ---------- TRANSAÇÃO ----------
    TransacaoRequest:
      type: object
      required: [tipo, descricao, valor]
      properties:
        tipo:
          type: string
          enum: [receita, despesa]
        descricao:
          type: string
        valor:
          type: number
          format: double
        ct_id:
          type: integer
          nullable: true
    Transacao:
      type: object
      properties:
        id:
          type: integer
        account_id:
          type: integer
        ct_id:
          type: integer
          nullable: true
        tipo:
          type: string
          enum: [receita, despesa]
        descricao:
          type: string
        valor:
          type: number
...
    AgendamentoRequest:
      type: object
      required: [aluno_id, horario_aula_id, data_aula]
      properties:
        aluno_id:
          type: integer
        horario_aula_id:
          type: integer
        data_aula:
          type: string
          format: date
        status:
          type: string
          enum: [agendado, cancelado, compareceu, faltou]
        observacao:
          type: string
          nullable: true
    Agendamento:
      type: object
      properties:
        id:
          type: integer
        account_id:
          type: integer
        aluno_id:
          type: integer
        horario_aula_id:
          type: integer
        data_aula:
          type: string
          format: date
        status:
          type: string
        observacao:
          type: string
          nullable: true

    # ---------- PRESENÇA ----------
    PresencaRequest:
      type: object
      required: [agendamento_id, status]
      properties:
        agendamento_id:
          type: integer
        status:
          type: string
        observacao:
          type: string
          nullable: true
    Presenca:
      type: object
      properties:
        id:
          type: integer
        account_id:
          type: integer
        agendamento_id:
          type: integer
        status:
          type: string
        observacao:
          type: string
          nullable: true

    # ---------- MENSALIDADE ----------
    MensalidadeRequest:
      type: object
      required: [aluno_id, competencia, valor, vencimento]
      properties:
        aluno_id:
          type: integer
        competencia:
          type: string
          pattern: '^\d{4}-(0[1-9]|1[0-2])$'
          example: '2026-04'
        valor:
          type: number
          format: double
        vencimento:
          type: string
          format: date
        observacao:
          type: string
          nullable: true
        data_pagamento:
          type: string
          format: date
          nullable: true
    Mensalidade:
      type: object
      properties:
        id:
          type: integer
        account_id:
          type: integer
        aluno_id:
          type: integer
        competencia:
          type: string
        valor:
          type: number
        vencimento:
          type: string
          format: date
        status:
          type: string
          enum: [pendente, pago, cancelado]
        data_pagamento:
          type: string
          format: date
          nullable: true
        observacao:
          type: string
          nullable: true
  parameters:
...
    Pagina:
      name: pagina
      in: query
      schema:
        type: integer
        default: 1
    Limite:
      name: limite
      in: query
      schema:
        type: integer
        default: 20
  responses:
...
    '401':
      description: Não autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'
...
    '409':
      description: Conflito (duplicidade)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'
security:
  - bearerAuth: []
paths:
...
  /cts:
    get:
      tags: [cts]
      summary: Listar CTs (paginado)
      parameters:
        - $ref: '#/components/parameters/Pagina'
        - $ref: '#/components/parameters/Limite'
      responses:
        '200':
          description: Lista de CTs
          content:
            application/json:
              schema:
                type: object
                properties:
                  dados:
                    type: array
                    items:
                      $ref: '#/components/schemas/Ct'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '401':
          $ref: '#/components/responses/401'
    post:
      tags: [cts]
      summary: Criar CT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CtRequest'
      responses:
        '201':
          description: CT criado
          content:
            application/json:
              schema:
                type: object
                properties:
                  dados:
                    type: object
                    properties:
                      mensagem:
                        type: string
                      id:
                        type: integer
        '409':
          $ref: '#/components/responses/409'
...
  /profissionais:
    get:
      tags: [profissionais]
      summary: Listar profissionais (paginado)
      parameters:
        - $ref: '#/components/parameters/Pagina'
        - $ref: '#/components/parameters/Limite'
      responses:
        '200':
          description: Lista de profissionais
          content:
            application/json:
              schema:
                type: object
                properties:
                  dados:
                    type: array
                    items:
                      $ref: '#/components/schemas/Profissional'
                  meta:
                    $ref: '#/components/schemas/Meta'
    post:
      tags: [profissionais]
      summary: Criar profissional
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfissionalRequest'
      responses:
        '201':
          description: Profissional criado
          content:
            application/json:
              schema:
                type: object
                properties:
                  dados:
                    type: object
                    properties:
                      mensagem:
                        type: string
                      id:
                        type: integer
        '409':
          $ref: '#/components/responses/409'
...
  /modalidades:
    get:
      tags: [modalidades]
      summary: Listar modalidades (paginado)
      parameters:
        - $ref: '#/components/parameters/Pagina'
        - $ref: '#/components/parameters/Limite'
      responses:
        '200':
          description: Lista de modalidades
          content:
            application/json:
              schema:
                type: object
                properties:
                  dados:
                    type: array
                    items:
                      $ref: '#/components/schemas/Modalidade'
                  meta:
                    $ref: '#/components/schemas/Meta'
    post:
      tags: [modalidades]
      summary: Criar modalidade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModalidadeRequest'
      responses:
        '201':
          description: Modalidade criada
          content:
            application/json:
              schema:
                type: object
                properties:
                  dados:
                    type: object
                    properties:
                      mensagem:
                        type: string
                      id:
                        type: integer
        '409':
          $ref: '#/components/responses/409'
...
  /horarios-aula:
    get:
      tags: [horarios-aula]
      summary: Listar horários (paginado)
      parameters:
        - $ref: '#/components/parameters/Pagina'
        - $ref: '#/components/parameters/Limite'
      responses:
        '200':
          description: Lista de horários
          content:
            application/json:
              schema:
                type: object
                properties:
                  dados:
                    type: array
                    items:
                      $ref: '#/components/schemas/HorarioAula'
                  meta:
                    $ref: '#/components/schemas/Meta'
    post:
      tags: [horarios-aula]
      summary: Criar horário de aula
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HorarioAulaRequest'
      responses:
        '201':
          description: Horário criado
          content:
            application/json:
              schema:
                type: object
                properties:
                  dados:
                    type: object
                    properties:
                      mensagem:
                        type: string
                      id:
                        type: integer
...
  /transacoes:
    get:
      tags: [transacoes]
      summary: Listar transações (paginado, com filtros)
      parameters:
        - $ref: '#/components/parameters/Pagina'
        - $ref: '#/components/parameters/Limite'
        - name: tipo
          in: query
          schema:
            type: string
            enum: [receita, despesa]
        - name: descricao
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Lista de transações
          content:
            application/json:
              schema:
                type: object
                properties:
                  dados:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transacao'
                  meta:
                    $ref: '#/components/schemas/Meta'
    post:
      tags: [transacoes]
      summary: Criar transação
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransacaoRequest'
      responses:
        '201':
          description: Transação criada
          content:
            application/json:
              schema:
                type: object
                properties:
                  dados:
                    type: object
                    properties:
                      mensagem:
                        type: string
                      id:
                        type: integer
        '409':
          $ref: '#/components/responses/409'
...
  /agendamentos:
    get:
      tags: [agendamentos]
      summary: Listar agendamentos (paginado)
      parameters:
        - $ref: '#/components/parameters/Pagina'
        - $ref: '#/components/parameters/Limite'
      responses:
        '200':
          description: Lista de agendamentos
          content:
            application/json:
              schema:
                type: object
                properties:
                  dados:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agendamento'
                  meta:
                    $ref: '#/components/schemas/Meta'
    post:
      tags: [agendamentos]
      summary: Criar agendamento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgendamentoRequest'
      responses:
        '201':
          description: Agendamento criado
          content:
            application/json:
              schema:
                type: object
                properties:
                  dados:
                    type: object
                    properties:
                      mensagem:
                        type: string
                      id:
                        type: integer
        '409':
          $ref: '#/components/responses/409'
...
  /presencas:
    get:
      tags: [presencas]
      summary: Listar presenças (paginado)
      parameters:
        - $ref: '#/components/parameters/Pagina'
        - $ref: '#/components/parameters/Limite'
      responses:
        '200':
          description: Lista de presenças
          content:
            application/json:
              schema:
                type: object
                properties:
                  dados:
                    type: array
                    items:
                      $ref: '#/components/schemas/Presenca'
                  meta:
                    $ref: '#/components/schemas/Meta'
    post:
      tags: [presencas]
      summary: Registrar presença
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresencaRequest'
      responses:
        '201':
          description: Presença registrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  dados:
                    type: object
                    properties:
                      mensagem:
                        type: string
                      id:
                        type: integer
        '409':
          $ref: '#/components/responses/409'
...
  /mensalidades:
    get:
      tags: [mensalidades]
      summary: Listar mensalidades (paginado, com filtros)
      parameters:
        - $ref: '#/components/parameters/Pagina'
        - $ref: '#/components/parameters/Limite'
        - name: aluno_id
          in: query
          schema:
            type: integer
        - name: competencia
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pendente, pago, cancelado]
      responses:
        '200':
          description: Lista de mensalidades
          content:
            application/json:
              schema:
                type: object
                properties:
                  dados:
                    type: array
                    items:
                      $ref: '#/components/schemas/Mensalidade'
                  meta:
                    $ref: '#/components/schemas/Meta'
    post:
      tags: [mensalidades]
      summary: Criar mensalidade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MensalidadeRequest'
      responses:
        '201':
          description: Mensalidade criada
          content:
            application/json:
              schema:
                type: object
                properties:
                  dados:
                    type: object
                    properties:
                      mensagem:
                        type: string
                      id:
                        type: integer
        '409':
          $ref: '#/components/responses/409'
...